move assign_commits to releases module
